requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==23.2.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.3.0
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.0
cycler==0.12.1
Cython==3.1.4
debugpy==1.8.15
decorator==4.4.2
defusedxml==0.7.1
EoN==1.2
executing==2.2.0
fastjsonschema==2.21.1
filelock>=3.15.4
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
juliacall==0.9.28
juliapkg==0.1.22
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
lark==1.2.2
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
multidict==6.0.5
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.0.0
optuna==4.4.0
overrides==7.7.0
packaging==24.1
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.33.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.26.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
sentry-sdk==2.43.0
setuptools==80.9.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.13.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.3
torch>=2.3.1
torch_geometric==2.5.3
torch_geometric_temporal==0.56.2
torch_scatter>=2.1.2
torch_sparse>=0.6.18
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.5.1
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.2.2
wandb==0.22.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
widgetsnbextension==4.0.14
yarl==1.9.4
